# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WhatsApp Auto-Reply Bot API
  description: |
    API for the WhatsApp Auto-Reply Bot SaaS platform.

    ## Authentication
    All protected endpoints require authentication via NextAuth.js session cookies.

    ## Rate Limiting
    - Login: 5 requests/minute
    - Registration: 3 requests/minute
    - Password Reset: 3 requests/minute
    - General API: 60 requests/minute

    ## Error Codes
    - 400: Bad Request (validation error)
    - 401: Unauthorized (not authenticated)
    - 403: Forbidden (insufficient permissions)
    - 404: Not Found
    - 429: Too Many Requests (rate limited)
    - 500: Internal Server Error
  version: 1.0.0
  contact:
    email: support@example.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
...
paths:
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 50
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                  description: Must contain uppercase letter and number
                locale:
                  type: string
                  enum: [ar, en]
                  default: ar
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/UserBasic'
                  requiresVerification:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already registered
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Password reset successful
        '400':
          description: Invalid or expired token
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /contacts:
    get:
      tags: [Contacts]
      summary: Get contacts list
      security:
        - sessionCookie: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: search
          in: query
          schema:
            type: string
            maxLength: 200
      responses:
        '200':
          description: Paginated contacts list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /messages:
    get:
      tags: [Messages]
      summary: Get messages list
      security:
        - sessionCookie: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: direction
          in: query
          schema:
            type: string
            enum: [INCOMING, OUTGOING]
        - name: contactId
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated messages list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: NextAuth.js session cookie
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unauthorized
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: object
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              retryAfter:
                type: integer
                description: Seconds until rate limit resets
  schemas:
    UserBasic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
...
    Contact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        phone:
          type: string
        name:
          type: string
          nullable: true
        messageCount:
          type: integer
        lastContact:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    ContactsResponse:
      type: object
      properties:
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/Contact'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        direction:
          type: string
          enum: [INCOMING, OUTGOING]
        status:
          type: string
          enum: [SENT, RECEIVED, FAILED]
        createdAt:
          type: string
          format: date-time
    MessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...